Rename toEmitNextValue to toEmitValue
