+++ src/link/http/__tests__/responseIterator.ts
@@ -189,7 +189,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -230,7 +230,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -257,7 +257,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -286,7 +286,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -315,7 +315,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -344,7 +344,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of batchedResults) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -382,7 +382,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
@@ -410,7 +410,7 @@ describe("multipart responses", () => {
     const observableStream = new ObservableStream(observable);
 
     for (const result of results) {
-      await expect(observableStream).toEmitNextValue(result);
+      await expect(observableStream).toEmitValue(result);
     }
 
     await expect(observableStream).toComplete();
