+++ src/testing/matchers/index.d.ts
@@ -48,7 +48,7 @@ interface ApolloCustomMatchers<R = void, T = {}> {
     (error?: any, options?: TakeOptions) => Promise<R>
   : { error: "matcher needs to be called on an ObservableStream instance" };
 
-  toEmitNextValue: T extends ObservableStream<any> ?
+  toEmitValue: T extends ObservableStream<any> ?
     (value: any, options?: TakeOptions) => Promise<R>
   : { error: "matcher needs to be called on an ObservableStream instance" };
 }
