+++ src/testing/matchers/index.ts
@@ -6,13 +6,13 @@ import { toBeDisposed } from "./toBeDisposed.js";
 import { toComplete } from "./toComplete.js";
 import { toEmitAnything } from "./toEmitAnything.js";
 import { toEmitError } from "./toEmitError.js";
-import { toEmitNextValue } from "./toEmitNextValue.js";
+import { toEmitValue } from "./toEmitValue.js";
 
 expect.extend({
   toComplete,
   toEmitAnything,
   toEmitError,
-  toEmitNextValue,
+  toEmitValue,
   toBeDisposed,
   toHaveSuspenseCacheEntryUsing,
   toMatchDocument,
